Fix more CodeRabbit issues
