+++ src/list-chats.tsx
@@ -3,34 +3,7 @@ import { ActionPanel, Detail, List, Action, Icon, Image } from "@raycast/api";
 import { withAccessToken } from "@raycast/utils";
 import { useBeeperDesktop, createBeeperOAuth, focusApp } from "./api";
 import { t } from "./locales";
-
-/**
- * Selects an icon image that represents a messaging network.
- *
- * @param network - The network name or identifier (case-insensitive; spaces, slashes, and dashes are ignored)
- * @returns The corresponding network icon image, or a generic message icon if the network is not recognized
- */
-function getNetworkIcon(network: string): Image.ImageLike {
-  const networkLower = network.toLowerCase().replace(/[/\s-]/g, "");
-
-  const iconMap: Record<string, string> = {
-    slack: "slack.svg",
-    whatsapp: "whatsapp.svg",
-    telegram: "telegram.svg",
-    discord: "discord.svg",
-    instagram: "instagram.svg",
-    facebook: "facebook.svg",
-    facebookmessenger: "messenger.svg",
-    messenger: "messenger.svg",
-    signal: "signal.svg",
-    imessage: "imessage.svg",
-    twitter: "twitter.svg",
-    email: "email.svg",
-    googlemessages: "google-messages.svg",
-  };
-
-  return iconMap[networkLower] || Icon.Message;
-}
+import { getNetworkIcon } from "./network-icons";
 
 /**
  * Render a searchable list of Beeper chats with actions to open the chat in Beeper, view details, and copy the chat ID.
@@ -67,9 +40,10 @@ function ListChatsCommand() {
           icon={getNetworkIcon(chat.network)}
           title={chat.title || translations.common.unnamedChat}
           subtitle={chat.network}
-          accessories={[{ text: chat.type }, chat.lastActivity ? { date: new Date(chat.lastActivity) } : {}].filter(
-            Boolean,
-          )}
+          accessories={[
+            { text: chat.type },
+            ...(chat.lastActivity ? [{ date: new Date(chat.lastActivity) }] : []),
+          ]}
           actions={
             <ActionPanel>
               <Action
@@ -82,7 +56,7 @@ function ListChatsCommand() {
                 icon={Icon.Info}
                 target={
                   <Detail
-                    markdown={`# ${chat.title}
+                    markdown={`# ${chat.title || translations.common.unnamedChat}
 
 **${translations.common.details.id}:** ${chat.id}
 **${translations.common.details.accountId}:** ${chat.accountID}
